pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "domarkx"
version = "0.1.0"
authors = [
    {name = "gameltb", email = "gamegccltb@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "autogen-agentchat",
    "autogen-core",
    "autogen-ext[openai]",
    "openai",
    "prompt-toolkit",
    "python-dotenv",
    "rich",
    "typer"
]
